refactor: edit app structure, edit addPlayerToRoom func
