+++ src/controllers/registration.ts
@@ -1,5 +1,5 @@
 import { db } from '../db';
-import { NewPlayer } from './player';
+import { NewPlayer } from '../db/player';
 import { MESSAGE_TYPES } from '../types/enums';
 import { WebSocketClient } from '../types/interfaces';
