+++ src/router/index.ts
@@ -6,11 +6,11 @@ import { WebSocketClient } from '../types/interfaces';
 
 export const router = (message: string, ws: WebSocketClient) => {
   try {
-    const data = JSON.parse(message);
+    const { type, data } = JSON.parse(message);
 
-    switch (data.type) {
+    switch (type) {
       case MESSAGE_TYPES.REGISTRATION:
-        const parsedData = JSON.parse(data.data);
+        const parsedData = JSON.parse(data);
         const { name, password } = parsedData;
         regPlayer(name, password, ws);
         break;
@@ -18,7 +18,8 @@ export const router = (message: string, ws: WebSocketClient) => {
         createRoom(ws);
         break;
       case MESSAGE_TYPES.ADD_USER_TO_ROOM:
-        addPlayerToRoom(data.data, ws);
+        const indexRoom = JSON.parse(data).indexRoom;
+        addPlayerToRoom(indexRoom, ws);
         break;
       default:
         console.log('Unknown message type');
